refactor: Use array with mapfile to read destination entries
feat: Remove entry if ssh-keyscan fails on target

The array refactor was necessary to avoid reading and writing to the
same file in the same pipeline. This is an alternative to doing a ping
test on all targets. If an ssh-keyscan fails, it's safe to assume that
the target is likely unreachable.